pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hackyeaster-solver"
version = "0.1.0"
description = "MCP server for Claude-driven HackyEaster CTF solving"
requires-python = ">=3.11"
dependencies = [
  "scrapling[fetchers]>=0.4.1",
  "lxml>=5.0",
  "mcp>=1.0.0",
  "Pillow>=10.0",
]

[project.scripts]
hackyeaster-mcp = "hackyeaster_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hackyeaster_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
